various options added for testing and presentation

+++ ThreadPoolManager.cpp
@@ -1,6 +1,6 @@
 #include"ThreadPoolManager.h"
 
-ThreadPoolManager::ThreadPoolManager(const size_t& work_thread_count, const size_t &output_thread_count, size_t& max_queue_size)
+ThreadPoolManager::ThreadPoolManager(const int& work_thread_count, const int&output_thread_count, int& max_queue_size)
 	: taskPool(new ThreadPool(work_thread_count, max_queue_size)), resultPool(new ThreadPool(output_thread_count))
 {
 }
